requirements.txt

Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-allauth==0.54.0
python-decouple==3.8
requests==2.31.0Django==5.1
django-cors-headers==3.14.0
numpy==1.24.3
torch==2.0.1